Move validateNetworks into common_openstacknet
It depends on GetOpenStackNetWithLabel and therefore makes sense
to keep it together to allign things for different api versions.
