+++ api/v1beta1/common_webhook.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 
 	"github.com/openstack-k8s-operators/osp-director-operator/api/shared"
-	k8s_errors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	goClient "sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -105,24 +104,4 @@ func IsUniqMAC(macNodeStatus map[string]OpenStackMACNodeReservation, mac string)
 		return true
 	}
 	return false
-}
-
-// validateNetworks - validate that for all configured subnets an osnet exists
-func validateNetworks(namespace string, networks []string) error {
-	for _, subnetName := range networks {
-		//
-		// Get OSnet with SubNetNameLabelSelector: subnetName
-		//
-		labelSelector := map[string]string{
-			shared.SubNetNameLabelSelector: subnetName,
-		}
-		osnet, err := GetOpenStackNetWithLabel(webhookClient, namespace, labelSelector)
-		if err != nil && k8s_errors.IsNotFound(err) {
-			return fmt.Errorf(fmt.Sprintf("%s %s not found, validate the object network list!", osnet.GetObjectKind().GroupVersionKind().Kind, subnetName))
-		} else if err != nil {
-			return fmt.Errorf(fmt.Sprintf("Failed to get %s %s", osnet.Kind, subnetName))
-		}
-	}
-
-	return nil
 }
